pyproject.toml
[tool.poetry]
name = "fast_user"
version = "0.1.0"
description = "The async-server for app user"
authors = ["Sergey Alekseev <alekseyserzh88@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.2"
sanic = "19.9.0"
python-dotenv = "0.10.3"
sanic-motor = "0.3.4"
sanic-openapi = "0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
